pyproject.toml
[project]
name = "youbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Tom Bedor", email = "tombedor@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.97.0",
    "gitpython>=3.1.47",
    "libtmux>=0.55.1",
    "pydantic>=2.13.3",
    "pyyaml>=6.0.3",
    "textual>=8.2.4",
]

[project.scripts]
youbot = "youbot:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "mypy>=1.20.2",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.12",
    "types-pyyaml>=6.0.12.20260408",
]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.mypy]
strict = true
python_version = "3.12"

[tool.pytest.ini_options]
asyncio_mode = "auto"
